Added Module Support to UI (#21)
In preparation for the calculator feature, the UI needed to not assume
that everything is a searcher or bookmarklet. This required switching
into dedicated 'modes' in the UI

Also batches a bug that relied on searching to always reset the window
size when the window toggled active. Now it's explicitly done, since
this side effect no longer occurs naturally.