Custom Styles support (#27)
Custom style support. The user can override the default style sheet in the dotfiles, which is created on launch if it does not exist.

Which required refactoring how window sizing is managed since that is now dynamic to the nature of the user stylesheet. This also meant removing support for centering placement, since it's again affected by the end user rendered sheet (as the window size changes after window creation but before fully rendered). This also meant making the app visible by default on launch, as resizing while the window is hidden during creation places it incorrectly...